Refactored analytics requests
