+++ .env.example
@@ -16,6 +16,7 @@ RECAPTCHA_PRIVATE_KEY=
 
 ANALYTICS_VIEW_ID=
 ANALYTICS_ENABLED=
+ANALYTICS_CACHE_STORE=file
 
 BROADCAST_DRIVER=log
 CACHE_DRIVER=file
