+++ README.md
@@ -15,6 +15,19 @@
 > ## Administration Preview
 > ![Admin](https://cloud.githubusercontent.com/assets/8210023/25923017/d958c432-35db-11e7-8306-92fc3406aed8.png)
 >
+> # Installation
+> ## Requirements
+>
+> |PHP|PHP Extension|DBMS|NodeJS|npm|Others (optional)
+> |---|---|---|---|---|---|
+> |![PHP](https://img.shields.io/badge/PHP->=7.0-0e7fbf.svg?style=flat-square)|![CURL](https://img.shields.io/badge/PHP%20ext-CURL-44CB12.svg?style=flat-square)<br>![OpenSSL](https://img.shields.io/badge/PHP%20ext-OpenSSL-44CB12.svg?style=flat-square)<br>![PDO](https://img.shields.io/badge/PHP%20ext-PDO-44CB12.svg?style=flat-square)<br>![Mbstring](https://img.shields.io/badge/PHP%20ext-Mbstring-44CB12.svg?style=flat-square)<br>![Tokenizer](https://img.shields.io/badge/PHP%20ext-Tokenizer-44CB12.svg?style=flat-square)<br>![XML](https://img.shields.io/badge/PHP%20ext-XML-44CB12.svg?style=flat-square)<br>![GD](https://img.shields.io/badge/PHP%20ext-GD-44CB12.svg?style=flat-square)|![MySQL](https://img.shields.io/badge/MySQL->=5.5-44CB12.svg?style=flat-square)|![NodeJS](https://img.shields.io/badge/NodeJS->=4-44CB12.svg?style=flat-square)|![npm](https://img.shields.io/badge/npm-*-44CB12.svg?style=flat-square)|![Analytics](https://img.shields.io/badge/Google-Analytics-44CB12.svg?style=flat-square)<br>![Redis](https://img.shields.io/badge/Redis-Server-44CB12.svg?style=flat-square)<br>![Redis](https://img.shields.io/badge/Redis-PHPRedis-44CB12.svg?style=flat-square)
+>
+> ## Install
+>
+> ```bash
+> composer create-project --prefer-dist xetaio/xetaravel <application_name>
+> ```
+>
 > ### Pre-installed Accounts
 > * **Admin**
 >   * User : **admin@xeta.io**
@@ -30,17 +43,5 @@
 >   * Password : **banished**
 >   * **Note** : You will need to delete the cookie to logout due to the restriction of the ban system.
 >
-> # Installation
-> ## Requirements
->
-> |PHP|PHP Extension|DBMS|NodeJS|npm
-> |---|---|---|---|---|
-> |![PHP](https://img.shields.io/badge/PHP->=7.0-44CB12.svg?style=flat-square)|![CURL](https://img.shields.io/badge/PHP%20ext-CURL-44CB12.svg?style=flat-square)<br>![OpenSSL](https://img.shields.io/badge/PHP%20ext-OpenSSL-44CB12.svg?style=flat-square)<br>![PDO](https://img.shields.io/badge/PHP%20ext-PDO-44CB12.svg?style=flat-square)<br>![Mbstring](https://img.shields.io/badge/PHP%20ext-Mbstring-44CB12.svg?style=flat-square)<br>![Tokenizer](https://img.shields.io/badge/PHP%20ext-Tokenizer-44CB12.svg?style=flat-square)<br>![XML](https://img.shields.io/badge/PHP%20ext-XML-44CB12.svg?style=flat-square)<br>![GD](https://img.shields.io/badge/PHP%20ext-GD-44CB12.svg?style=flat-square)|![MySQL](https://img.shields.io/badge/MySQL->=5.5-44CB12.svg?style=flat-square)|![NodeJS](https://img.shields.io/badge/NodeJS->=4-44CB12.svg?style=flat-square)|![npm](https://img.shields.io/badge/npm-*-44CB12.svg?style=flat-square)
->
-> ## Install
->
-> ```bash
-> composer create-project --prefer-dist xetaio/xetaravel <application_name>
-> ```
 > # Contribute
 > If you want to contribute to the project by adding new features or just fix a bug, feel free to do a PR.
