+++ config/database.php
@@ -92,8 +92,8 @@ return [
     */
 
     'redis' => [
-
         'cluster' => false,
+        'client' => 'phpredis',
 
         'default' => [
             'host' => env('REDIS_HOST', '127.0.0.1'),
