+++ config/session.php
@@ -70,7 +70,7 @@ return [
     |
     */
 
-    'connection' => null,
+    'connection' => 'default',
 
     /*
     |--------------------------------------------------------------------------
